feat: 🎸 Add identities namespace
Creates an identities namespace and moves methods into it

- sdk.registerIdentity -> sdk.identities.registerIdentity
- sdk.getIdentity -> sdk.identities.getIdentitiy
- sdk.isIdentityValid -> sdk.identities.isIdentityValid

BREAKING CHANGE: 🧨 Moves Identity methods
