+++ src/Polymesh.ts
@@ -13,6 +13,7 @@ import { satisfies } from 'semver';
 import { w3cwebsocket as W3CWebSocket } from 'websocket';
 import WebSocketAsPromised from 'websocket-as-promised';
 
+import { Identities } from '~/Identities';
 import {
   Account,
   claimClassicTicker,
@@ -20,8 +21,6 @@ import {
   Context,
   Identity,
   PolymeshError,
-  registerIdentity,
-  RegisterIdentityParams,
   SecurityToken,
   TickerReservation,
   transferPolyx,
@@ -82,6 +81,10 @@ export class Polymesh {
   public middleware: Middleware;
   public settlements: Settlements;
   public currentIdentity: CurrentIdentity;
+  /**
+   * A set of methods for interacting with Polymesh Identities.
+   */
+  public identities: Identities;
 
   /**
    * @hidden
@@ -95,14 +98,10 @@ export class Polymesh {
     this.middleware = new Middleware(context);
     this.settlements = new Settlements(context);
     this.currentIdentity = new CurrentIdentity(context);
+    this.identities = new Identities(context);
 
     this.transferPolyx = createProcedureMethod(
       { getProcedureAndArgs: args => [transferPolyx, args] },
-      context
-    );
-
-    this.registerIdentity = createProcedureMethod(
-      { getProcedureAndArgs: args => [registerIdentity, args] },
       context
     );
 
@@ -458,15 +457,6 @@ export class Polymesh {
   }
 
   /**
-   * Create an Identity instance from a DID
-   *
-   * @throws if there is no Identity with the passed DID
-   */
-  public async getIdentity(args: { did: string }): Promise<Identity> {
-    return this.context.getIdentity(args.did);
-  }
-
-  /**
    * Retrieve the Identity associated to the current Account (null if there is none)
    */
   public getCurrentIdentity(): Promise<Identity | null> {
@@ -493,16 +483,6 @@ export class Polymesh {
    */
   public getAccounts(): Account[] {
     return this.context.getAccounts();
-  }
-
-  /**
-   * Return whether the supplied Identity/DID exists
-   */
-  public async isIdentityValid(args: { identity: Identity | string }): Promise<boolean> {
-    const { identity: did } = args;
-    const identity = did instanceof Identity ? did : new Identity({ did }, this.context);
-
-    return identity.exists();
   }
 
   /**
@@ -659,19 +639,6 @@ export class Polymesh {
     const { free } = await account.getBalance();
     return free;
   }
-
-  /**
-   * Register an Identity
-   *
-   * @note must be a CDD provider
-   * @note this may create [[AuthorizationRequest | Authorization Requests]] which have to be accepted by
-   *   the corresponding [[Account | Accounts]] and/or [[Identity | Identities]]. An Account or Identity can
-   *   fetch its pending Authorization Requests by calling `authorizations.getReceived`
-   *
-   * @note required role:
-   *   - Customer Due Diligence Provider
-   */
-  public registerIdentity: ProcedureMethod<RegisterIdentityParams, Identity>;
 
   /**
    * Retrieve the number of the latest block in the chain
